Open source the tests.
Running the file run_tests.py will run the tests. By default, it only runs a subset of tests, because running all the tests takes too long.

Unfortunately, the tests spew a huge amount of output to the terminal. This is useful for debugging failed tests, but it's hard to see what output is for which test.

PiperOrigin-RevId: 185591518